Silence B009/B010 for non-identifiers
If an object implements __getattr__ and/or __setattr__, the second
parameter to getattr() or setattr() may not be a valid Python
identifier (e.g. getattr(foo, "123abc"). In this case bugbear should
not emit B009/B010, since changing to "foo.123abc" would be a
SyntaxError. Update the B009/B010 detection to check using a regex.